+++ test/unit/classes/resultset.cpp
@@ -880,7 +880,7 @@ void resultset::getResultSetType()
     connection_properties.erase("defaultStatementResultType");
     {
       logMsg("... testing defaultStatementResultType");
-      connection_properties["defaultStatementResultType"]=(long long)sql::ResultSet::TYPE_FORWARD_ONLY;
+      connection_properties["defaultStatementResultType"]= sql::ResultSet::TYPE_FORWARD_ONLY;
       try
       {
         created_objects.clear();
@@ -899,7 +899,7 @@ void resultset::getResultSetType()
       ASSERT_EQUALS(pstmt->getResultSetType(), sql::ResultSet::TYPE_SCROLL_INSENSITIVE);
 
       connection_properties.erase("defaultStatementResultType");
-      connection_properties["defaultStatementResultType"]=(long long)sql::ResultSet::TYPE_SCROLL_INSENSITIVE;
+      connection_properties["defaultStatementResultType"]= sql::ResultSet::TYPE_SCROLL_INSENSITIVE;
       try
       {
         created_objects.clear();
@@ -916,7 +916,7 @@ void resultset::getResultSetType()
       ASSERT_EQUALS(pstmt->getResultSetType(), sql::ResultSet::TYPE_SCROLL_INSENSITIVE);
 
       connection_properties.erase("defaultStatementResultType");
-      connection_properties["defaultStatementResultType"]=(long long)sql::ResultSet::TYPE_SCROLL_SENSITIVE;
+      connection_properties["defaultStatementResultType"]= sql::ResultSet::TYPE_SCROLL_SENSITIVE;
       try
       {
         created_objects.clear();
